Fixed revert errors. Updated demo for proper contrast multiplier mapping. Added test to generateBaseScale.
