--- packages/base-scale/package.json
@@ -1,51 +1,0 @@
-{
-  "name": "@adobe/leonardo-base-scale",
-  "version": "1.0.0-alpha.1",
-  "description": "Generate base scale using HSLuv value distribution",
-  "repository": "git@github.com:adobe/leonardo.git",
-  "type": "module",
-  "main": "./dist/index.js",
-  "module": "./index.js",
-  "exports": {
-    ".": {
-      "require": "./dist/index.js",
-      "default": "./index.js"
-    }
-  },
-  "scripts": {
-    "build": "babel *.js --out-dir dist",
-    "prepublishOnly": "yarn build",
-    "dev": "jest --watch",
-    "test": "jest"
-  },
-  "author": "Nate Baldwin <nbaldwin@adobe.com>",
-  "license": "Apache-2.0",
-  "devDependencies": {
-    "@babel/cli": "^7.7.5",
-    "@babel/core": "^7.7.5",
-    "@babel/preset-env": "^7.7.6",
-    "jest": "^24.9.0"
-  },
-  "dependencies": {
-    "d3": "^5.14.2",
-    "d3-3d": "0.0.9",
-    "d3-cam02": "^0.1.5",
-    "d3-hsluv": "^0.1.2",
-    "d3-hsv": "^0.1.0"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "babel": {
-    "presets": [
-      [
-        "@babel/preset-env",
-        {
-          "targets": {
-            "node": "current"
-          }
-        }
-      ]
-    ]
-  }
-}
