+++ packages/contrast-colors/test/generateBaseScale.test.js
@@ -32,3 +32,10 @@ test('should generate 101 blue toned grayscale colors in LAB space)', function()
   expect(colors).toEqual([ '#000000', '#020202', '#050608', '#080a0d', '#0b0d10', '#0e1013', '#111317', '#131519', '#14171c', '#16191e', '#181b21', '#191d24', '#1b1f27', '#1c2129', '#1e232c', '#20252f', '#212732', '#232935', '#252b39', '#272e3c', '#28303e', '#2a3241', '#2b3444', '#2e3647', '#2f394b', '#313b4e', '#333d50', '#343f54', '#364257', '#38445a', '#39465d', '#3c4861', '#3d4b64', '#3f4d67', '#414f6b', '#43526e', '#445471', '#475675', '#485978', '#4b5c7b', '#4d5e7d', '#4f607f', '#526382', '#546583', '#576886', '#596a88', '#5c6d8a', '#5e6f8c', '#61728e', '#647490', '#677793', '#697995', '#6c7c97', '#6f7f99', '#71819b', '#74849d', '#7786a0', '#7989a2', '#7c8ca4', '#7f8ea6', '#8291a9', '#8493ab', '#8796ad', '#8a98b0', '#8d9bb2', '#909eb4', '#93a0b6', '#95a3b9', '#99a6bb', '#9ca8be', '#9eabc0', '#a1aec2', '#a4b0c4', '#a7b3c7', '#aab6c9', '#aeb9cb', '#b0bbcd', '#b4becf', '#b7c0d1', '#bac3d3', '#bdc6d5', '#c0c9d7', '#c4ccd9', '#c6cedb', '#cad1dd', '#cdd4df', '#d1d7e1', '#d3d9e3', '#d6dce5', '#dadfe7', '#dde2e9', '#e1e5ec', '#e3e7ed', '#e7eaef', '#eaedf1', '#edf0f4', '#f1f2f6', '#f4f5f8', '#f7f8fa', '#fafbfc', '#fdfefe' ]);
 
 });
+
+test('should generate 101 blue colors in HSLuv space)', function() {
+  let colors = generateBaseScale({colorKeys: ['#6FA7FF', '#B5E6FF'], colorspace: 'HSLuv'});;
+
+  expect(colors).toEqual([ '#000001', '#000207', '#000612', '#000a1b', '#000d20', '#001025', '#00132a', '#00152e', '#001731', '#001935', '#001b38', '#001d3b', '#001f3f', '#002143', '#002347', '#00254b', '#00284e', '#002a52', '#002c55', '#002e59', '#00305c', '#003260', '#003465', '#003768', '#00396c', '#003b6f', '#003d74', '#004078', '#00427b', '#00447f', '#004784', '#004987', '#004b8b', '#004d90', '#005094', '#005297', '#00559c', '#00579f', '#0059a4', '#005ca8', '#005eac', '#0061b1', '#0063b4', '#0066b9', '#0068bd', '#006ac2', '#006dc5', '#0070cb', '#0072ce', '#0074d3', '#0077d7', '#0079dc', '#007ce0', '#007fe4', '#0081e8', '#0084ed', '#0086f2', '#0089f6', '#008cfb', '#008efe', '#2191ff', '#2f93ff', '#3b96ff', '#4799ff', '#4e9bff', '#579eff', '#5da0ff', '#64a3ff', '#6ca6ff', '#6fa9ff', '#70acff', '#71b0ff', '#73b3ff', '#75b6ff', '#77baff', '#79bdff', '#7cc0ff', '#7fc3ff', '#82c7ff', '#86caff', '#8acdff', '#8ed0ff', '#92d2ff', '#96d5ff', '#9ad8ff', '#9fdbff', '#a4deff', '#aae1ff', '#aee3ff', '#b4e6ff', '#bbe8ff', '#c1eaff', '#c9ecff', '#d1efff', '#d7f1ff', '#ddf3ff', '#e3f5ff', '#eaf7ff', '#f1faff', '#f6fcff', '#fcfeff' ]);
+
+});
